Extract `ViolationHash` class
I've been studying the RuboCop engine as I work on an engine for Reek.

There's a lot happening in `rubocop.rb`, with many mixed levels of
abstraction. It seemed that extracting out the responsibility of
building the violation hash would make things easier to understand and
maintain.

I've not yet updated the tests to reflect the new structure. I'll await
feedback on this proposal first.
